# refactor: reduce the number of deps for public libraries