--- a/core/primitives-core/src/hash.rs
+++ b/core/primitives-core/src/hash.rs
@@ -7,12 +7,16 @@ use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use crate::logging::pretty_hash;
 use crate::serialize::{from_base, to_base, BaseDecode};
 
-#[derive(Copy, Clone, PartialOrd, PartialEq, Eq, Ord, derive_more::AsRef)]
-#[as_ref(forward)]
+#[derive(Copy, Clone, PartialOrd, PartialEq, Eq, Ord)]
 pub struct Digest(pub [u8; 32]);
 
-#[derive(Copy, Clone, PartialOrd, Ord, derive_more::AsRef)]
-#[as_ref(forward)]
+impl AsRef<[u8]> for Digest {
+    fn as_ref(&self) -> &[u8] {
+        self.0.as_ref()
+    }
+}
+
+#[derive(Copy, Clone, PartialOrd, Ord)]
 pub struct CryptoHash(pub Digest);
 
 impl<'a> From<&'a CryptoHash> for String {
@@ -24,6 +28,12 @@ impl<'a> From<&'a CryptoHash> for String {
 impl Default for CryptoHash {
     fn default() -> Self {
         CryptoHash(Digest(Default::default()))
+    }
+}
+
+impl AsRef<[u8]> for CryptoHash {
+    fn as_ref(&self) -> &[u8] {
+        self.0.as_ref()
     }
 }