primitives-core is an API crate, so all of its deps become deps of
consumers of the library. This makes deps significantly more costly.

For derive-more in particular, it seems like it is not used that
heavily.