@matklad 2021-04-01
@frol this is more of a suggestion PR: I don't have enough experience to understand whether we want to do this or not. I think you are the better person to make a judgement call here :)