@matklad 2021-04-02
> our crates granularity is correct before we try to save some bits.

That's a great point, agree that this is premature optimization at this point, closing!

> measure the improvement

I've tried `cd core/primitives-core && t cargo build --lib -j 1`, the results look like this:

```rust
# With derive more
real 32.10s
cpu  33.02s (30.78s user + 2.24s sys)
rss  452.62mb

# Without derive more
real 28.18s
cpu  28.93s (26.89s user + 2.04s sys)
rss  392.70mb
```

So that's 4 seconds absolute / 14 % relative difference in *the worst-case* (from scratch debug build). `derive-more` seems to be among the bigger ones in this corener of the crate graps:

![image](https://user-images.githubusercontent.com/1711539/113409060-2beb3600-93b9-11eb-8215-487e86034cc1.png)

(`env RUSTC_BOOTSTRAP=1 cargo build --lib -j 1 -Ztimings`)

> and enforce it

The enforcement that works well for rust-analayzer is three-fold:

* dependency stance [is documented](https://github.com/rust-analyzer/rust-analyzer/blob/master/docs/dev/style.md#cratesio-dependencies)
* by the time folks get review permissions, they generally know that we don't do deps
* once in a while I read Cargo.lock to see if I need to raise some issues for our dependencies :)