@matklad 2021-04-02
Forgot to mention one more thing: 

>My measurement is that with derive_more I need to read less code

I personally feel the opposite: with derive_more, I need to read *more* code. It's not immediately obvious if it implements `AsRef<[u8]>` or `AsRef<[u8; 32]>`, and if I "goto definition", than I land into proc macro, which is not too transparent.