@frol 2021-04-02
Great points! I would love to see those measures applied. Currently, we only use `cargo-deny` to protect us from too many duplicate dependencies (we [cannot avoid them completely since](https://github.com/near/nearcore/blob/master/deny.toml) we pull out quite massive third-party dependencies like wasmer, wasmtime, evm, and various crypto implementations)